Fix bugs
* Fix CDummy ctor

* Fix enum usage CWeapon::FireSniper

* Fix CTrafficLights::ShouldCarStopForBridge

* Fix CRGBA::operator*=

* Fix CFileLoader::LoadAnimatedClumpObject

* Fix CEventEditableResponse::InformRespectedFriends

* Fix CAEAudioChannel::SetFrequencyScalingFactor

* Fix MarkSurroundingEntitiesForCollisionWithTrain

* FIX_BUGS CPopulation::ConvertToDummyObject

* FIX_BUGS CTaskComplexUseGoggles::ControlSubTask

* Remove unnecessary check WaterCreature::Exit

* Remove copy constructor CWeapon

* Remove debug leftovers CollisionDebugModule

* Refactor CTaskComplexUseSequence::Clone

* Remove hooks CModelInfoAccelerator

+++ source/game_sa/Collision/CollisionData.h
@@ -105,7 +105,7 @@ public:
 
 private:
     // HELPERS
-    template <typename T> T* GetPointerToColArray(uint32 byteOffset) {
+    template <typename T> T* GetPointerToColArray(size_t byteOffset) {
         return reinterpret_cast<T*>(&reinterpret_cast<uint8*>(this)[byteOffset]);
     }
